split of stdlib.h for each port
